과제 진행 중 퇴근으로 인한 백업
[todo]
- post 인피니티 스크롤 구현 중...

+++ spring-rest-api/src/main/java/org/restapi/springrestapi/service/comment/CommentServiceImpl.java
@@ -7,35 +7,51 @@ import org.restapi.springrestapi.dto.comment.CommentResult;
 import org.restapi.springrestapi.dto.comment.PatchCommentRequest;
 import org.restapi.springrestapi.dto.comment.RegisterCommentRequest;
 import org.restapi.springrestapi.finder.CommentFinder;
+import org.restapi.springrestapi.finder.PostFinder;
 import org.restapi.springrestapi.model.Comment;
-import org.restapi.springrestapi.legacy.repository.CommentRepository;
+import org.restapi.springrestapi.repository.CommentRepository;
+import org.restapi.springrestapi.model.Post;
+import org.restapi.springrestapi.repository.PostRepository;
+import org.restapi.springrestapi.validator.CommentValidator;
+import org.restapi.springrestapi.validator.PostValidator;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
-@RequestMapping
 @RequiredArgsConstructor
+@Transactional
 public class CommentServiceImpl implements CommentService {
 	private final CommentRepository commentRepository;
+    private final PostRepository postRepository;
+
 	private final CommentFinder commentFinder;
 	private final CommentValidator commentValidator;
+    private final PostFinder postFinder;
+    private final PostValidator postValidator;
 
-	@Override
+    @Override
 	public CommentResult registerComment(Long userId, RegisterCommentRequest request, Long postId) {
 		commentValidator.validateOrigin(userId, postId);
 
-		Comment comment = Comment.from(request, userId, postId);
+        Post post = postFinder.findProxyById(postId);
+		Comment comment = Comment.from(request, post);
+        postRepository.increaseCommentCount(postId);
 
 		return CommentResult.from(commentRepository.save(comment));
-
 	}
 
 	@Override
 	public CommentListResult getCommentList(Long postId, int cursor, int limit) {
-		commentValidator.validatePost(postId);
+		postValidator.validatePostExists(postId);
+
 		List<CommentResult> commentList = commentFinder.findAll(postId, cursor, limit);
+
+		if (commentList.isEmpty()) {
+			return CommentListResult.empty();
+		}
+
 		return CommentListResult.from(commentList, (int)Math.max(commentList.get(0).id() - 1, 0));
 
 	}
@@ -43,16 +59,20 @@ public class CommentServiceImpl implements CommentService {
 	@Override
 	public CommentResult updateComment(Long userId, PatchCommentRequest request, Long postId, Long id) {
 		commentValidator.validateOrigin(userId, postId);
-		commentValidator.validateComment(id);
+        commentValidator.validateOwner(userId, id);
+
 		Comment comment = commentFinder.findById(id);
 		comment.updateContent(request.content());
+
 		return CommentResult.from(commentRepository.save(comment));
 	}
 
 	@Override
 	public void deleteComment(Long userId, Long id, Long postId) {
 		commentValidator.validateOrigin(userId, postId);
-		commentValidator.validateComment(id);
+        commentValidator.validateOwner(userId, id);
+
+        postRepository.decreaseCommentCount(postId);
 		commentRepository.deleteById(id);
 	}
 }
